refactor: 统一路径的接口形式

+++ src/backend/api/file/delete.rs
@@ -1,37 +1,31 @@
+use axum::Json;
 use serde::Deserialize;
 use tokio::fs;
 
-use crate::backend::{
-  db::DBConnection,
-  error::AppError,
-  extractor::storage::{StorageExtractor, WithStorage},
-};
+use crate::backend::{db::DBConnection, error::AppError, extractor::storage::Storage};
 
 #[derive(Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct DeleteFileDto {
-  name: String,
-  path: String,
-}
-
-impl WithStorage for DeleteFileDto {
-  fn get_path(&self) -> &str {
-    &self.path
-  }
+  targets: Vec<String>,
 }
 
 #[axum::debug_handler(state = DBConnection)]
 pub async fn delete_file(
-  StorageExtractor(dto, local_path): StorageExtractor<DeleteFileDto>,
+  Storage(local_path): Storage,
+  Json(dto): Json<DeleteFileDto>,
 ) -> Result<(), AppError> {
-  let name = dto.name;
-  let local_path = local_path.safe_join(&name)?;
-  log::info!("delete_file: {}", local_path.display());
-
-  if !local_path.exists() {
-    return Err(AppError::new("文件不存在"));
+  for target in dto.targets {
+    let local_path = local_path.safe_join(&target)?;
+    if !local_path.exists() {
+      log::error!("file not found: {}", local_path.display());
+      continue;
+    }
+    if local_path.is_dir() {
+      log::error!("file is a directory: {}", local_path.display());
+      continue;
+    }
+    fs::remove_file(&local_path).await?;
   }
-
-  fs::remove_file(&local_path).await?;
   Ok(())
 }
